% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abrir_base.R
\name{abrir_base}
\alias{abrir_base}
\title{FUNCTION_TITLE}
\usage{
abrir_base(
  base,
  estados = NULL,
  censo_dir = "~/Downloads/Censo2010",
  cores = 1
)
}
\arguments{
\item{base}{PARAM_DESCRIPTION}

\item{estados}{PARAM_DESCRIPTION, Default: NULL}

\item{censo_dir}{PARAM_DESCRIPTION, Default: '~/Downloads/Censo2010'}

\item{cores}{PARAM_DESCRIPTION, Default: 1}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[foreach]{foreach}}
}
