% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abrir_base_estado.R
\name{abrir_base_estado}
\alias{abrir_base_estado}
\title{FUNCTION_TITLE}
\usage{
abrir_base_estado(base, estado, censo_dir = "~/Downloads/Censo2010/")
}
\arguments{
\item{base}{PARAM_DESCRIPTION}

\item{estado}{PARAM_DESCRIPTION}

\item{censo_dir}{PARAM_DESCRIPTION, Default: '~/Downloads/Censo2010/'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[readxl]{read_excel}}
}
