% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basico.R
\docType{data}
\name{Basico}
\alias{Basico}
\title{DATASET_TITLE}
\format{
A data frame with 310120 rows and 33 variables:
\describe{
  \item{\code{Cod_setor}}{double COLUMN_DESCRIPTION}
  \item{\code{Cod_Grandes Regiões}}{integer COLUMN_DESCRIPTION}
  \item{\code{Nome_Grande_Regiao}}{character COLUMN_DESCRIPTION}
  \item{\code{Cod_UF}}{integer COLUMN_DESCRIPTION}
  \item{\code{Nome_da_UF}}{character COLUMN_DESCRIPTION}
  \item{\code{Cod_meso}}{integer COLUMN_DESCRIPTION}
  \item{\code{Nome_da_meso}}{character COLUMN_DESCRIPTION}
  \item{\code{Cod_micro}}{integer COLUMN_DESCRIPTION}
  \item{\code{Nome_da_micro}}{character COLUMN_DESCRIPTION}
  \item{\code{Cod_RM}}{integer COLUMN_DESCRIPTION}
  \item{\code{Nome_da_RM}}{character COLUMN_DESCRIPTION}
  \item{\code{Cod_municipio}}{integer COLUMN_DESCRIPTION}
  \item{\code{Nome_do_municipio}}{character COLUMN_DESCRIPTION}
  \item{\code{Cod_distrito}}{integer COLUMN_DESCRIPTION}
  \item{\code{Nome_do_distrito}}{character COLUMN_DESCRIPTION}
  \item{\code{Cod_subdistrito}}{double COLUMN_DESCRIPTION}
  \item{\code{Nome_do_subdistrito}}{character COLUMN_DESCRIPTION}
  \item{\code{Cod_bairro}}{integer COLUMN_DESCRIPTION}
  \item{\code{Nome_do_bairro}}{character COLUMN_DESCRIPTION}
  \item{\code{Situacao_setor}}{integer COLUMN_DESCRIPTION}
  \item{\code{Tipo_setor}}{double COLUMN_DESCRIPTION}
  \item{\code{V001}}{double COLUMN_DESCRIPTION}
  \item{\code{V002}}{double COLUMN_DESCRIPTION}
  \item{\code{V003}}{double COLUMN_DESCRIPTION}
  \item{\code{V004}}{double COLUMN_DESCRIPTION}
  \item{\code{V005}}{double COLUMN_DESCRIPTION}
  \item{\code{V006}}{double COLUMN_DESCRIPTION}
  \item{\code{V007}}{double COLUMN_DESCRIPTION}
  \item{\code{V008}}{double COLUMN_DESCRIPTION}
  \item{\code{V009}}{double COLUMN_DESCRIPTION}
  \item{\code{V010}}{double COLUMN_DESCRIPTION}
  \item{\code{V011}}{double COLUMN_DESCRIPTION}
  \item{\code{V012}}{double COLUMN_DESCRIPTION}
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
Basico
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
