% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censo2010brasil-package.R
\docType{package}
\name{censo2010brasil-package}
\alias{censo2010brasil}
\alias{censo2010brasil-package}
\title{censo2010brasil: Dados do Censo Demográfico 2010 do Brasil}
\description{
Bases de agregados por setores censitários (Censo Brasil 2010)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
