% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DomicilioRenda.R
\docType{data}
\name{DomicilioRenda}
\alias{DomicilioRenda}
\title{DATASET_TITLE}
\format{
A data frame with 310120 rows and 15 variables:
\describe{
  \item{\code{Cod_setor}}{double COLUMN_DESCRIPTION}
  \item{\code{V001}}{double COLUMN_DESCRIPTION}
  \item{\code{V002}}{double COLUMN_DESCRIPTION}
  \item{\code{V003}}{double COLUMN_DESCRIPTION}
  \item{\code{V004}}{double COLUMN_DESCRIPTION}
  \item{\code{V005}}{double COLUMN_DESCRIPTION}
  \item{\code{V006}}{double COLUMN_DESCRIPTION}
  \item{\code{V007}}{double COLUMN_DESCRIPTION}
  \item{\code{V008}}{double COLUMN_DESCRIPTION}
  \item{\code{V009}}{double COLUMN_DESCRIPTION}
  \item{\code{V010}}{double COLUMN_DESCRIPTION}
  \item{\code{V011}}{double COLUMN_DESCRIPTION}
  \item{\code{V012}}{double COLUMN_DESCRIPTION}
  \item{\code{V013}}{double COLUMN_DESCRIPTION}
  \item{\code{V014}}{double COLUMN_DESCRIPTION}
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
DomicilioRenda
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
