% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Entorno04.R
\docType{data}
\name{Entorno04}
\alias{Entorno04}
\title{DATASET_TITLE}
\format{
A data frame with 0 rows and 241 variables:
\describe{
  \item{\code{Cod_setor}}{double COLUMN_DESCRIPTION}
  \item{\code{Situacao_setor}}{double COLUMN_DESCRIPTION}
  \item{\code{V623}}{double COLUMN_DESCRIPTION}
  \item{\code{V624}}{double COLUMN_DESCRIPTION}
  \item{\code{V625}}{double COLUMN_DESCRIPTION}
  \item{\code{V626}}{double COLUMN_DESCRIPTION}
  \item{\code{V627}}{double COLUMN_DESCRIPTION}
  \item{\code{V628}}{double COLUMN_DESCRIPTION}
  \item{\code{V629}}{double COLUMN_DESCRIPTION}
  \item{\code{V630}}{double COLUMN_DESCRIPTION}
  \item{\code{V631}}{double COLUMN_DESCRIPTION}
  \item{\code{V632}}{double COLUMN_DESCRIPTION}
  \item{\code{V633}}{double COLUMN_DESCRIPTION}
  \item{\code{V634}}{double COLUMN_DESCRIPTION}
  \item{\code{V635}}{double COLUMN_DESCRIPTION}
  \item{\code{V636}}{double COLUMN_DESCRIPTION}
  \item{\code{V637}}{double COLUMN_DESCRIPTION}
  \item{\code{V638}}{double COLUMN_DESCRIPTION}
  \item{\code{V639}}{double COLUMN_DESCRIPTION}
  \item{\code{V640}}{double COLUMN_DESCRIPTION}
  \item{\code{V641}}{double COLUMN_DESCRIPTION}
  \item{\code{V642}}{double COLUMN_DESCRIPTION}
  \item{\code{V643}}{double COLUMN_DESCRIPTION}
  \item{\code{V644}}{double COLUMN_DESCRIPTION}
  \item{\code{V645}}{double COLUMN_DESCRIPTION}
  \item{\code{V646}}{double COLUMN_DESCRIPTION}
  \item{\code{V647}}{double COLUMN_DESCRIPTION}
  \item{\code{V648}}{double COLUMN_DESCRIPTION}
  \item{\code{V649}}{double COLUMN_DESCRIPTION}
  \item{\code{V650}}{double COLUMN_DESCRIPTION}
  \item{\code{V651}}{double COLUMN_DESCRIPTION}
  \item{\code{V652}}{double COLUMN_DESCRIPTION}
  \item{\code{V653}}{double COLUMN_DESCRIPTION}
  \item{\code{V654}}{double COLUMN_DESCRIPTION}
  \item{\code{V655}}{double COLUMN_DESCRIPTION}
  \item{\code{V656}}{double COLUMN_DESCRIPTION}
  \item{\code{V657}}{double COLUMN_DESCRIPTION}
  \item{\code{V658}}{double COLUMN_DESCRIPTION}
  \item{\code{V659}}{double COLUMN_DESCRIPTION}
  \item{\code{V660}}{double COLUMN_DESCRIPTION}
  \item{\code{V661}}{double COLUMN_DESCRIPTION}
  \item{\code{V662}}{double COLUMN_DESCRIPTION}
  \item{\code{V663}}{double COLUMN_DESCRIPTION}
  \item{\code{V664}}{double COLUMN_DESCRIPTION}
  \item{\code{V665}}{double COLUMN_DESCRIPTION}
  \item{\code{V666}}{double COLUMN_DESCRIPTION}
  \item{\code{V667}}{double COLUMN_DESCRIPTION}
  \item{\code{V668}}{double COLUMN_DESCRIPTION}
  \item{\code{V669}}{double COLUMN_DESCRIPTION}
  \item{\code{V670}}{double COLUMN_DESCRIPTION}
  \item{\code{V671}}{double COLUMN_DESCRIPTION}
  \item{\code{V672}}{double COLUMN_DESCRIPTION}
  \item{\code{V673}}{double COLUMN_DESCRIPTION}
  \item{\code{V674}}{double COLUMN_DESCRIPTION}
  \item{\code{V675}}{double COLUMN_DESCRIPTION}
  \item{\code{V676}}{double COLUMN_DESCRIPTION}
  \item{\code{V677}}{double COLUMN_DESCRIPTION}
  \item{\code{V678}}{double COLUMN_DESCRIPTION}
  \item{\code{V679}}{double COLUMN_DESCRIPTION}
  \item{\code{V680}}{double COLUMN_DESCRIPTION}
  \item{\code{V681}}{double COLUMN_DESCRIPTION}
  \item{\code{V682}}{double COLUMN_DESCRIPTION}
  \item{\code{V683}}{double COLUMN_DESCRIPTION}
  \item{\code{V684}}{double COLUMN_DESCRIPTION}
  \item{\code{V685}}{double COLUMN_DESCRIPTION}
  \item{\code{V686}}{double COLUMN_DESCRIPTION}
  \item{\code{V687}}{double COLUMN_DESCRIPTION}
  \item{\code{V688}}{double COLUMN_DESCRIPTION}
  \item{\code{V689}}{double COLUMN_DESCRIPTION}
  \item{\code{V690}}{double COLUMN_DESCRIPTION}
  \item{\code{V691}}{double COLUMN_DESCRIPTION}
  \item{\code{V692}}{double COLUMN_DESCRIPTION}
  \item{\code{V693}}{double COLUMN_DESCRIPTION}
  \item{\code{V694}}{double COLUMN_DESCRIPTION}
  \item{\code{V695}}{double COLUMN_DESCRIPTION}
  \item{\code{V696}}{double COLUMN_DESCRIPTION}
  \item{\code{V697}}{double COLUMN_DESCRIPTION}
  \item{\code{V698}}{double COLUMN_DESCRIPTION}
  \item{\code{V699}}{double COLUMN_DESCRIPTION}
  \item{\code{V700}}{double COLUMN_DESCRIPTION}
  \item{\code{V701}}{double COLUMN_DESCRIPTION}
  \item{\code{V702}}{double COLUMN_DESCRIPTION}
  \item{\code{V703}}{double COLUMN_DESCRIPTION}
  \item{\code{V704}}{double COLUMN_DESCRIPTION}
  \item{\code{V705}}{double COLUMN_DESCRIPTION}
  \item{\code{V706}}{double COLUMN_DESCRIPTION}
  \item{\code{V707}}{double COLUMN_DESCRIPTION}
  \item{\code{V708}}{double COLUMN_DESCRIPTION}
  \item{\code{V709}}{double COLUMN_DESCRIPTION}
  \item{\code{V710}}{double COLUMN_DESCRIPTION}
  \item{\code{V711}}{double COLUMN_DESCRIPTION}
  \item{\code{V712}}{double COLUMN_DESCRIPTION}
  \item{\code{V713}}{double COLUMN_DESCRIPTION}
  \item{\code{V714}}{double COLUMN_DESCRIPTION}
  \item{\code{V715}}{double COLUMN_DESCRIPTION}
  \item{\code{V716}}{double COLUMN_DESCRIPTION}
  \item{\code{V717}}{double COLUMN_DESCRIPTION}
  \item{\code{V718}}{double COLUMN_DESCRIPTION}
  \item{\code{V719}}{double COLUMN_DESCRIPTION}
  \item{\code{V720}}{double COLUMN_DESCRIPTION}
  \item{\code{V721}}{double COLUMN_DESCRIPTION}
  \item{\code{V722}}{double COLUMN_DESCRIPTION}
  \item{\code{V723}}{double COLUMN_DESCRIPTION}
  \item{\code{V724}}{double COLUMN_DESCRIPTION}
  \item{\code{V725}}{double COLUMN_DESCRIPTION}
  \item{\code{V726}}{double COLUMN_DESCRIPTION}
  \item{\code{V727}}{double COLUMN_DESCRIPTION}
  \item{\code{V728}}{double COLUMN_DESCRIPTION}
  \item{\code{V729}}{double COLUMN_DESCRIPTION}
  \item{\code{V730}}{double COLUMN_DESCRIPTION}
  \item{\code{V731}}{double COLUMN_DESCRIPTION}
  \item{\code{V732}}{double COLUMN_DESCRIPTION}
  \item{\code{V733}}{double COLUMN_DESCRIPTION}
  \item{\code{V734}}{double COLUMN_DESCRIPTION}
  \item{\code{V735}}{double COLUMN_DESCRIPTION}
  \item{\code{V736}}{double COLUMN_DESCRIPTION}
  \item{\code{V737}}{double COLUMN_DESCRIPTION}
  \item{\code{V738}}{double COLUMN_DESCRIPTION}
  \item{\code{V739}}{double COLUMN_DESCRIPTION}
  \item{\code{V740}}{double COLUMN_DESCRIPTION}
  \item{\code{V741}}{double COLUMN_DESCRIPTION}
  \item{\code{V742}}{double COLUMN_DESCRIPTION}
  \item{\code{V743}}{double COLUMN_DESCRIPTION}
  \item{\code{V744}}{double COLUMN_DESCRIPTION}
  \item{\code{V745}}{double COLUMN_DESCRIPTION}
  \item{\code{V746}}{double COLUMN_DESCRIPTION}
  \item{\code{V747}}{double COLUMN_DESCRIPTION}
  \item{\code{V748}}{double COLUMN_DESCRIPTION}
  \item{\code{V749}}{double COLUMN_DESCRIPTION}
  \item{\code{V750}}{double COLUMN_DESCRIPTION}
  \item{\code{V751}}{double COLUMN_DESCRIPTION}
  \item{\code{V752}}{double COLUMN_DESCRIPTION}
  \item{\code{V753}}{double COLUMN_DESCRIPTION}
  \item{\code{V754}}{double COLUMN_DESCRIPTION}
  \item{\code{V755}}{double COLUMN_DESCRIPTION}
  \item{\code{V756}}{double COLUMN_DESCRIPTION}
  \item{\code{V757}}{double COLUMN_DESCRIPTION}
  \item{\code{V758}}{double COLUMN_DESCRIPTION}
  \item{\code{V759}}{double COLUMN_DESCRIPTION}
  \item{\code{V760}}{double COLUMN_DESCRIPTION}
  \item{\code{V761}}{double COLUMN_DESCRIPTION}
  \item{\code{V762}}{double COLUMN_DESCRIPTION}
  \item{\code{V763}}{double COLUMN_DESCRIPTION}
  \item{\code{V764}}{double COLUMN_DESCRIPTION}
  \item{\code{V765}}{double COLUMN_DESCRIPTION}
  \item{\code{V766}}{double COLUMN_DESCRIPTION}
  \item{\code{V767}}{double COLUMN_DESCRIPTION}
  \item{\code{V768}}{double COLUMN_DESCRIPTION}
  \item{\code{V769}}{double COLUMN_DESCRIPTION}
  \item{\code{V770}}{double COLUMN_DESCRIPTION}
  \item{\code{V771}}{double COLUMN_DESCRIPTION}
  \item{\code{V772}}{double COLUMN_DESCRIPTION}
  \item{\code{V773}}{double COLUMN_DESCRIPTION}
  \item{\code{V774}}{double COLUMN_DESCRIPTION}
  \item{\code{V775}}{double COLUMN_DESCRIPTION}
  \item{\code{V776}}{double COLUMN_DESCRIPTION}
  \item{\code{V777}}{double COLUMN_DESCRIPTION}
  \item{\code{V778}}{double COLUMN_DESCRIPTION}
  \item{\code{V779}}{double COLUMN_DESCRIPTION}
  \item{\code{V780}}{double COLUMN_DESCRIPTION}
  \item{\code{V781}}{double COLUMN_DESCRIPTION}
  \item{\code{V782}}{double COLUMN_DESCRIPTION}
  \item{\code{V783}}{double COLUMN_DESCRIPTION}
  \item{\code{V784}}{double COLUMN_DESCRIPTION}
  \item{\code{V785}}{double COLUMN_DESCRIPTION}
  \item{\code{V786}}{double COLUMN_DESCRIPTION}
  \item{\code{V787}}{double COLUMN_DESCRIPTION}
  \item{\code{V788}}{double COLUMN_DESCRIPTION}
  \item{\code{V789}}{double COLUMN_DESCRIPTION}
  \item{\code{V790}}{double COLUMN_DESCRIPTION}
  \item{\code{V791}}{double COLUMN_DESCRIPTION}
  \item{\code{V792}}{double COLUMN_DESCRIPTION}
  \item{\code{V793}}{double COLUMN_DESCRIPTION}
  \item{\code{V794}}{double COLUMN_DESCRIPTION}
  \item{\code{V795}}{double COLUMN_DESCRIPTION}
  \item{\code{V796}}{double COLUMN_DESCRIPTION}
  \item{\code{V797}}{double COLUMN_DESCRIPTION}
  \item{\code{V798}}{double COLUMN_DESCRIPTION}
  \item{\code{V799}}{double COLUMN_DESCRIPTION}
  \item{\code{V800}}{double COLUMN_DESCRIPTION}
  \item{\code{V801}}{double COLUMN_DESCRIPTION}
  \item{\code{V802}}{double COLUMN_DESCRIPTION}
  \item{\code{V803}}{double COLUMN_DESCRIPTION}
  \item{\code{V804}}{double COLUMN_DESCRIPTION}
  \item{\code{V805}}{double COLUMN_DESCRIPTION}
  \item{\code{V806}}{double COLUMN_DESCRIPTION}
  \item{\code{V807}}{double COLUMN_DESCRIPTION}
  \item{\code{V808}}{double COLUMN_DESCRIPTION}
  \item{\code{V809}}{double COLUMN_DESCRIPTION}
  \item{\code{V810}}{double COLUMN_DESCRIPTION}
  \item{\code{V811}}{double COLUMN_DESCRIPTION}
  \item{\code{V812}}{double COLUMN_DESCRIPTION}
  \item{\code{V813}}{double COLUMN_DESCRIPTION}
  \item{\code{V814}}{double COLUMN_DESCRIPTION}
  \item{\code{V815}}{double COLUMN_DESCRIPTION}
  \item{\code{V816}}{double COLUMN_DESCRIPTION}
  \item{\code{V817}}{double COLUMN_DESCRIPTION}
  \item{\code{V818}}{double COLUMN_DESCRIPTION}
  \item{\code{V819}}{double COLUMN_DESCRIPTION}
  \item{\code{V820}}{double COLUMN_DESCRIPTION}
  \item{\code{V821}}{double COLUMN_DESCRIPTION}
  \item{\code{V822}}{double COLUMN_DESCRIPTION}
  \item{\code{V823}}{double COLUMN_DESCRIPTION}
  \item{\code{V824}}{double COLUMN_DESCRIPTION}
  \item{\code{V825}}{double COLUMN_DESCRIPTION}
  \item{\code{V826}}{double COLUMN_DESCRIPTION}
  \item{\code{V827}}{double COLUMN_DESCRIPTION}
  \item{\code{V828}}{double COLUMN_DESCRIPTION}
  \item{\code{V829}}{double COLUMN_DESCRIPTION}
  \item{\code{V830}}{double COLUMN_DESCRIPTION}
  \item{\code{V831}}{double COLUMN_DESCRIPTION}
  \item{\code{V832}}{double COLUMN_DESCRIPTION}
  \item{\code{V833}}{double COLUMN_DESCRIPTION}
  \item{\code{V834}}{double COLUMN_DESCRIPTION}
  \item{\code{V835}}{double COLUMN_DESCRIPTION}
  \item{\code{V836}}{double COLUMN_DESCRIPTION}
  \item{\code{V837}}{double COLUMN_DESCRIPTION}
  \item{\code{V838}}{double COLUMN_DESCRIPTION}
  \item{\code{V839}}{double COLUMN_DESCRIPTION}
  \item{\code{V840}}{double COLUMN_DESCRIPTION}
  \item{\code{V841}}{double COLUMN_DESCRIPTION}
  \item{\code{V842}}{double COLUMN_DESCRIPTION}
  \item{\code{Setor_Precoleta}}{double COLUMN_DESCRIPTION}
  \item{\code{Cod_Grandes Regiões}}{double COLUMN_DESCRIPTION}
  \item{\code{Nome_Grande_Regiao}}{character COLUMN_DESCRIPTION}
  \item{\code{Cod_UF}}{double COLUMN_DESCRIPTION}
  \item{\code{Nome_da_UF}}{character COLUMN_DESCRIPTION}
  \item{\code{Cod_meso}}{double COLUMN_DESCRIPTION}
  \item{\code{Nome_da_meso}}{character COLUMN_DESCRIPTION}
  \item{\code{Cod_micro}}{double COLUMN_DESCRIPTION}
  \item{\code{Nome_da_micro}}{character COLUMN_DESCRIPTION}
  \item{\code{Cod_RM}}{double COLUMN_DESCRIPTION}
  \item{\code{Nome_da_RM}}{character COLUMN_DESCRIPTION}
  \item{\code{Cod_municipio}}{double COLUMN_DESCRIPTION}
  \item{\code{Nome_do_municipio}}{character COLUMN_DESCRIPTION}
  \item{\code{Cod_distrito}}{double COLUMN_DESCRIPTION}
  \item{\code{Nome_do_distrito}}{character COLUMN_DESCRIPTION}
  \item{\code{Cod_subdistrito}}{double COLUMN_DESCRIPTION}
  \item{\code{Nome_do_subdistrito}}{character COLUMN_DESCRIPTION}
  \item{\code{Cod_bairro}}{double COLUMN_DESCRIPTION}
  \item{\code{Nome_do_bairro}}{character COLUMN_DESCRIPTION}
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
Entorno04
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
