% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pessoa02.R
\docType{data}
\name{Pessoa02}
\alias{Pessoa02}
\title{DATASET_TITLE}
\format{
A data frame with 310120 rows and 171 variables:
\describe{
  \item{\code{Cod_setor}}{double COLUMN_DESCRIPTION}
  \item{\code{V086}}{double COLUMN_DESCRIPTION}
  \item{\code{V087}}{double COLUMN_DESCRIPTION}
  \item{\code{V088}}{double COLUMN_DESCRIPTION}
  \item{\code{V089}}{double COLUMN_DESCRIPTION}
  \item{\code{V090}}{double COLUMN_DESCRIPTION}
  \item{\code{V091}}{double COLUMN_DESCRIPTION}
  \item{\code{V092}}{double COLUMN_DESCRIPTION}
  \item{\code{V093}}{double COLUMN_DESCRIPTION}
  \item{\code{V094}}{double COLUMN_DESCRIPTION}
  \item{\code{V095}}{double COLUMN_DESCRIPTION}
  \item{\code{V096}}{double COLUMN_DESCRIPTION}
  \item{\code{V097}}{double COLUMN_DESCRIPTION}
  \item{\code{V098}}{double COLUMN_DESCRIPTION}
  \item{\code{V099}}{double COLUMN_DESCRIPTION}
  \item{\code{V100}}{double COLUMN_DESCRIPTION}
  \item{\code{V101}}{double COLUMN_DESCRIPTION}
  \item{\code{V102}}{double COLUMN_DESCRIPTION}
  \item{\code{V103}}{double COLUMN_DESCRIPTION}
  \item{\code{V104}}{double COLUMN_DESCRIPTION}
  \item{\code{V105}}{double COLUMN_DESCRIPTION}
  \item{\code{V106}}{double COLUMN_DESCRIPTION}
  \item{\code{V107}}{double COLUMN_DESCRIPTION}
  \item{\code{V108}}{double COLUMN_DESCRIPTION}
  \item{\code{V109}}{double COLUMN_DESCRIPTION}
  \item{\code{V110}}{double COLUMN_DESCRIPTION}
  \item{\code{V111}}{double COLUMN_DESCRIPTION}
  \item{\code{V112}}{double COLUMN_DESCRIPTION}
  \item{\code{V113}}{double COLUMN_DESCRIPTION}
  \item{\code{V114}}{double COLUMN_DESCRIPTION}
  \item{\code{V115}}{double COLUMN_DESCRIPTION}
  \item{\code{V116}}{double COLUMN_DESCRIPTION}
  \item{\code{V117}}{double COLUMN_DESCRIPTION}
  \item{\code{V118}}{double COLUMN_DESCRIPTION}
  \item{\code{V119}}{double COLUMN_DESCRIPTION}
  \item{\code{V120}}{double COLUMN_DESCRIPTION}
  \item{\code{V121}}{double COLUMN_DESCRIPTION}
  \item{\code{V122}}{double COLUMN_DESCRIPTION}
  \item{\code{V123}}{double COLUMN_DESCRIPTION}
  \item{\code{V124}}{double COLUMN_DESCRIPTION}
  \item{\code{V125}}{double COLUMN_DESCRIPTION}
  \item{\code{V126}}{double COLUMN_DESCRIPTION}
  \item{\code{V127}}{double COLUMN_DESCRIPTION}
  \item{\code{V128}}{double COLUMN_DESCRIPTION}
  \item{\code{V129}}{double COLUMN_DESCRIPTION}
  \item{\code{V130}}{double COLUMN_DESCRIPTION}
  \item{\code{V131}}{double COLUMN_DESCRIPTION}
  \item{\code{V132}}{double COLUMN_DESCRIPTION}
  \item{\code{V133}}{double COLUMN_DESCRIPTION}
  \item{\code{V134}}{double COLUMN_DESCRIPTION}
  \item{\code{V135}}{double COLUMN_DESCRIPTION}
  \item{\code{V136}}{double COLUMN_DESCRIPTION}
  \item{\code{V137}}{double COLUMN_DESCRIPTION}
  \item{\code{V138}}{double COLUMN_DESCRIPTION}
  \item{\code{V139}}{double COLUMN_DESCRIPTION}
  \item{\code{V140}}{double COLUMN_DESCRIPTION}
  \item{\code{V141}}{double COLUMN_DESCRIPTION}
  \item{\code{V142}}{double COLUMN_DESCRIPTION}
  \item{\code{V143}}{double COLUMN_DESCRIPTION}
  \item{\code{V144}}{double COLUMN_DESCRIPTION}
  \item{\code{V145}}{double COLUMN_DESCRIPTION}
  \item{\code{V146}}{double COLUMN_DESCRIPTION}
  \item{\code{V147}}{double COLUMN_DESCRIPTION}
  \item{\code{V148}}{double COLUMN_DESCRIPTION}
  \item{\code{V149}}{double COLUMN_DESCRIPTION}
  \item{\code{V150}}{double COLUMN_DESCRIPTION}
  \item{\code{V151}}{double COLUMN_DESCRIPTION}
  \item{\code{V152}}{double COLUMN_DESCRIPTION}
  \item{\code{V153}}{double COLUMN_DESCRIPTION}
  \item{\code{V154}}{double COLUMN_DESCRIPTION}
  \item{\code{V155}}{double COLUMN_DESCRIPTION}
  \item{\code{V156}}{double COLUMN_DESCRIPTION}
  \item{\code{V157}}{double COLUMN_DESCRIPTION}
  \item{\code{V158}}{double COLUMN_DESCRIPTION}
  \item{\code{V159}}{double COLUMN_DESCRIPTION}
  \item{\code{V160}}{double COLUMN_DESCRIPTION}
  \item{\code{V161}}{double COLUMN_DESCRIPTION}
  \item{\code{V162}}{double COLUMN_DESCRIPTION}
  \item{\code{V163}}{double COLUMN_DESCRIPTION}
  \item{\code{V164}}{double COLUMN_DESCRIPTION}
  \item{\code{V165}}{double COLUMN_DESCRIPTION}
  \item{\code{V166}}{double COLUMN_DESCRIPTION}
  \item{\code{V167}}{double COLUMN_DESCRIPTION}
  \item{\code{V168}}{double COLUMN_DESCRIPTION}
  \item{\code{V169}}{double COLUMN_DESCRIPTION}
  \item{\code{V170}}{double COLUMN_DESCRIPTION}
  \item{\code{V171}}{double COLUMN_DESCRIPTION}
  \item{\code{V172}}{double COLUMN_DESCRIPTION}
  \item{\code{V173}}{double COLUMN_DESCRIPTION}
  \item{\code{V174}}{double COLUMN_DESCRIPTION}
  \item{\code{V175}}{double COLUMN_DESCRIPTION}
  \item{\code{V176}}{double COLUMN_DESCRIPTION}
  \item{\code{V177}}{double COLUMN_DESCRIPTION}
  \item{\code{V178}}{double COLUMN_DESCRIPTION}
  \item{\code{V179}}{double COLUMN_DESCRIPTION}
  \item{\code{V180}}{double COLUMN_DESCRIPTION}
  \item{\code{V181}}{double COLUMN_DESCRIPTION}
  \item{\code{V182}}{double COLUMN_DESCRIPTION}
  \item{\code{V183}}{double COLUMN_DESCRIPTION}
  \item{\code{V184}}{double COLUMN_DESCRIPTION}
  \item{\code{V185}}{double COLUMN_DESCRIPTION}
  \item{\code{V186}}{double COLUMN_DESCRIPTION}
  \item{\code{V187}}{double COLUMN_DESCRIPTION}
  \item{\code{V188}}{double COLUMN_DESCRIPTION}
  \item{\code{V189}}{double COLUMN_DESCRIPTION}
  \item{\code{V190}}{double COLUMN_DESCRIPTION}
  \item{\code{V191}}{double COLUMN_DESCRIPTION}
  \item{\code{V192}}{double COLUMN_DESCRIPTION}
  \item{\code{V193}}{double COLUMN_DESCRIPTION}
  \item{\code{V194}}{double COLUMN_DESCRIPTION}
  \item{\code{V195}}{double COLUMN_DESCRIPTION}
  \item{\code{V196}}{double COLUMN_DESCRIPTION}
  \item{\code{V197}}{double COLUMN_DESCRIPTION}
  \item{\code{V198}}{double COLUMN_DESCRIPTION}
  \item{\code{V199}}{double COLUMN_DESCRIPTION}
  \item{\code{V200}}{double COLUMN_DESCRIPTION}
  \item{\code{V201}}{double COLUMN_DESCRIPTION}
  \item{\code{V202}}{double COLUMN_DESCRIPTION}
  \item{\code{V203}}{double COLUMN_DESCRIPTION}
  \item{\code{V204}}{double COLUMN_DESCRIPTION}
  \item{\code{V205}}{double COLUMN_DESCRIPTION}
  \item{\code{V206}}{double COLUMN_DESCRIPTION}
  \item{\code{V207}}{double COLUMN_DESCRIPTION}
  \item{\code{V208}}{double COLUMN_DESCRIPTION}
  \item{\code{V209}}{double COLUMN_DESCRIPTION}
  \item{\code{V210}}{double COLUMN_DESCRIPTION}
  \item{\code{V211}}{double COLUMN_DESCRIPTION}
  \item{\code{V212}}{double COLUMN_DESCRIPTION}
  \item{\code{V213}}{double COLUMN_DESCRIPTION}
  \item{\code{V214}}{double COLUMN_DESCRIPTION}
  \item{\code{V215}}{double COLUMN_DESCRIPTION}
  \item{\code{V216}}{double COLUMN_DESCRIPTION}
  \item{\code{V217}}{double COLUMN_DESCRIPTION}
  \item{\code{V218}}{double COLUMN_DESCRIPTION}
  \item{\code{V219}}{double COLUMN_DESCRIPTION}
  \item{\code{V220}}{double COLUMN_DESCRIPTION}
  \item{\code{V221}}{double COLUMN_DESCRIPTION}
  \item{\code{V222}}{double COLUMN_DESCRIPTION}
  \item{\code{V223}}{double COLUMN_DESCRIPTION}
  \item{\code{V224}}{double COLUMN_DESCRIPTION}
  \item{\code{V225}}{double COLUMN_DESCRIPTION}
  \item{\code{V226}}{double COLUMN_DESCRIPTION}
  \item{\code{V227}}{double COLUMN_DESCRIPTION}
  \item{\code{V228}}{double COLUMN_DESCRIPTION}
  \item{\code{V229}}{double COLUMN_DESCRIPTION}
  \item{\code{V230}}{double COLUMN_DESCRIPTION}
  \item{\code{V231}}{double COLUMN_DESCRIPTION}
  \item{\code{V232}}{double COLUMN_DESCRIPTION}
  \item{\code{V233}}{double COLUMN_DESCRIPTION}
  \item{\code{V234}}{double COLUMN_DESCRIPTION}
  \item{\code{V235}}{double COLUMN_DESCRIPTION}
  \item{\code{V236}}{double COLUMN_DESCRIPTION}
  \item{\code{V237}}{double COLUMN_DESCRIPTION}
  \item{\code{V238}}{double COLUMN_DESCRIPTION}
  \item{\code{V239}}{double COLUMN_DESCRIPTION}
  \item{\code{V240}}{double COLUMN_DESCRIPTION}
  \item{\code{V241}}{double COLUMN_DESCRIPTION}
  \item{\code{V242}}{double COLUMN_DESCRIPTION}
  \item{\code{V243}}{double COLUMN_DESCRIPTION}
  \item{\code{V244}}{double COLUMN_DESCRIPTION}
  \item{\code{V245}}{double COLUMN_DESCRIPTION}
  \item{\code{V246}}{double COLUMN_DESCRIPTION}
  \item{\code{V247}}{double COLUMN_DESCRIPTION}
  \item{\code{V248}}{double COLUMN_DESCRIPTION}
  \item{\code{V249}}{double COLUMN_DESCRIPTION}
  \item{\code{V250}}{double COLUMN_DESCRIPTION}
  \item{\code{V251}}{double COLUMN_DESCRIPTION}
  \item{\code{V252}}{double COLUMN_DESCRIPTION}
  \item{\code{V253}}{double COLUMN_DESCRIPTION}
  \item{\code{V254}}{double COLUMN_DESCRIPTION}
  \item{\code{V255}}{double COLUMN_DESCRIPTION}
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
Pessoa02
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
