% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pessoa10.R
\docType{data}
\name{Pessoa10}
\alias{Pessoa10}
\title{DATASET_TITLE}
\format{
A data frame with 0 rows and 5 variables:
\describe{
  \item{\code{Cod_setor}}{double COLUMN_DESCRIPTION}
  \item{\code{Situacao_setor}}{double COLUMN_DESCRIPTION}
  \item{\code{V001}}{double COLUMN_DESCRIPTION}
  \item{\code{V002}}{double COLUMN_DESCRIPTION}
  \item{\code{V003}}{double COLUMN_DESCRIPTION}
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
Pessoa10
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
